feat: add SubmitProposalButton component for proposal submission with transaction handling
